fix SIZE_T definition to ctypes.c_size_t + fix MessageAttribute._get_required_buffer_size

+++ ctypes_generation/generate.py
@@ -22,7 +22,7 @@ TYPE_EQUIVALENCE = [
     ('BYTE', 'c_ubyte'),
     ('PWSTR', 'LPWSTR'),
     ('PCWSTR', 'LPWSTR'),
-    ('SIZE_T', 'c_ulong'),
+    ('SIZE_T', 'c_size_t'),
     ('PSIZE_T', 'POINTER(SIZE_T)'),
     ('PVOID', 'c_void_p'),
     ('PPS_POST_PROCESS_INIT_ROUTINE', 'PVOID'),
